refactor: draw the GPU component with *iced*

+++ src/config.rs
@@ -74,7 +74,7 @@ pub struct DiskProps {
 
 #[derive(Deserialize)]
 pub struct GpuProps {
-  pub update_interval: u32,
+  pub update_interval: u64,
 }
 
 #[derive(Deserialize)]
